chore(api): split tests
